requirements.txt
pyrogram
python-dotenv
fpdf
moviepy
gTTS
